Changed all priced to BigDecimal from double

+++ Capstone-Chris-main/ecommerce/src/main/java/com/hcl/ecommerce/entity/CartItem.java
@@ -1,5 +1,7 @@
 package com.hcl.ecommerce.entity;
 
+import java.math.BigDecimal;
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -35,7 +37,7 @@ public class CartItem {
 	@Column(nullable = false)
 	private int quantity;
 	
-	private double subtotal;
+	private BigDecimal subtotal;
 	
 	@ManyToOne
 	@JoinColumn(name = "user_id", nullable = false)
@@ -78,11 +80,11 @@ public class CartItem {
 		this.quantity = quantity;
 	}
 
-	public double getSubtotal() {
+	public BigDecimal getSubtotal() {
 		return subtotal;
 	}
 
-	public void setSubtotal(double subtotal) {
+	public void setSubtotal(BigDecimal subtotal) {
 		this.subtotal = subtotal;
 	}
